# [LTD-3562] Optimise case detail endpoint