--- a/Pipfile
+++ b/Pipfile
@@ -73,6 +73,7 @@ celery = "~=5.2.7"
 redis = "~=4.0.2"
 psycopg2-binary = "~=2.9.3"
 django-test-migrations = "~=1.2.0"
+django-silk = "==5.0.3"
 
 [requires]
 python_version = "3.7"