--- a/api/applications/serializers/standard_application.py
+++ b/api/applications/serializers/standard_application.py
@@ -117,7 +117,7 @@ class StandardApplicationViewSerializer(PartiesSerializerMixin, GenericApplicati
         ]
 
     def get_denial_matches(self, instance):
-        denial_matches = DenialMatchOnApplication.objects.filter(application=instance, denial__is_revoked=False)
+        denial_matches = instance.denial_matches.filter(denial__is_revoked=False)
         return DenialMatchOnApplicationViewSerializer(denial_matches, many=True).data
 
     def get_is_amended(self, instance):