--- a/api/cases/views/views.py
+++ b/api/cases/views/views.py
@@ -118,9 +118,19 @@ class CaseDetail(APIView):
                 "advice__countersigned_by__team",
                 "advice__countersigned_by__role",
                 "advice__denial_reasons",
+                "advice__good",
+                "advice__good__good",
+                "countersign_advice",
+                "flags",
+                "queues",
+                "copy_of",
+                "copy_of__status",
+                "copy_of__status__status",
             ],
             select_related=[
                 "case_type",
+                "case_officer",
+                "case_officer__team",
             ],
         )
         data = CaseDetailSerializer(case, user=gov_user, team=gov_user.team).data