--- a/api/conf/urls.py
+++ b/api/conf/urls.py
@@ -47,3 +47,6 @@ if settings.ADMIN_ENABLED:
             path("admin/login/", api.core.views.LoginProviderView.as_view()),
             path("auth/", include("authbroker_client.urls")),
         ] + urlpatterns
+
+if settings.ENABLE_DJANGO_SILK:
+    urlpatterns += [path("silk/", include("silk.urls", namespace="silk"))]