### Aim
Changes here speed up the case detail page.  Broadly the changes made were as follows;
- Add django-silk behind the env var `ENABLE_DJANGO_SILK` - for profiling slow requests.
- Add a bunch more prefetch related calls in order to reduce the number of queries made by the GET case endpoint.
- Reduce additional queries by using django ORM relationships for lookups (populated with prefetch/select related) instead of whole new queries.
- Remove `good_application_documents`/`good_application_internal_documents` fields from `GoodOnApplicationViewSerializer` as the frontend does not use these fields and they were costly to populate.  These were maintained for Data workspace endpoints through using an extended serializer.
- Remove `precedents` field from `GoodSerializerInternal` as this is not used in the frontend application (other precedents endpoints are used instead).  This allowed a large speed increase for certain troublesome cases which had a lot of precedents.  Again, this is maintained for data workspace endpoints to avoid changing the contract radically there.

This combination of changes (along with the frontend changes https://github.com/uktrade/lite-frontend/pull/1229) have allowed the following speed increases for specific case detail pages on my local working copy;
```
Local dev copy results

http://localhost:8200/queues/00000000-0000-0000-0000-000000000001/cases/5064d020-134e-43d3-ad51-00bc742e1209/details/
Previous; 2600ms
Now; 900ms
~3X speed increase

http://localhost:8200/queues/4dfc0d0b-c0c0-4f59-897b-b7d596bed930/cases/ad7c9771-f76b-40ad-90e2-336ba5fa913a/
Previous; 10000ms
Now; 3700ms
~3X speed increase

http://localhost:8200/queues/00000000-0000-0000-0000-000000000001/cases/ca0d0afc-29ef-493f-9365-0eaefb8ea69f/
Previous; 5400ms
Now; 1400ms
~4X speed increase

http://localhost:8200/queues/00000000-0000-0000-0000-000000000001/cases/c39c55b3-aade-4256-a941-c27072260f7f/
Previous; 18000ms
Now; 2000ms
~9X speed increase

http://localhost:8200/queues/00000000-0000-0000-0000-000000000001/cases/bbc1d33c-b26c-4486-9e21-0caf841d7ad8/
Previous; 2500ms
Now; 908ms
~3X speed increase

http://localhost:8200/queues/00000000-0000-0000-0000-000000000001/cases/da289c38-383c-4298-aaa5-8d27987276f8/
Previous; 2400ms
Now; 970ms
~2.5X speed increase
```

Note that we may not see exactly the same level of speed increases in production (due to network/hardware/latency variables).

[LTD-3562](https://uktrade.atlassian.net/browse/LTD-3562)


[LTD-3562]: https://uktrade.atlassian.net/browse/LTD-3562?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ